more

+++ Bets4You/Bets4You/ComboCoef.cs
@@ -8,17 +8,17 @@ namespace Bets4You
     public class ComboCoef
     {
         private List<int> betIndexes;
-        private double sumCoef;
+        private int sumCoef;
         private string submitorName;
 
-        public ComboCoef(List<int> betindex, double sumcoef, string submitorName)
+        public ComboCoef(List<int> betindex, int sumcoef, string submitorName)
         {
             this.betIndexes = betindex;
             this.sumCoef = sumcoef;
             this.submitorName = submitorName;
         }
         public List<int> BetIndex { get => betIndexes; set => betIndexes = value; }
-        public double SumCoeff { get => sumCoef; set => sumCoef = value; }
+        public int SumCoeff { get => sumCoef; set => sumCoef = value; }
         public string SubmitorName { get => submitorName; set => submitorName = value; }
     }
 }
